새로운 내용으로 덮어쓰기

+++ src/main/java/kr/sparta/lunchvote/dto/round/LunchRoundResponseDto.java
@@ -19,20 +19,20 @@ public class LunchRoundResponseDto {
     private List<MenuResponseDto> menus;
     private int totalVotes;
 
-    // 생성자: LunchRound + 로그인 사용자 + VoteRepository
     public LunchRoundResponseDto(LunchRound round, User user, VoteRepository voteRepository) {
         this.id = round.getId();
         this.userName = round.getUser().getName();
         this.date = round.getDate() != null ? round.getDate().toString() : null;
 
-        // MenuResponseDto 변환 시 voteCount와 votedByMe 계산
+
         this.menus = round.getMenus().stream()
                 .map(menu -> {
-                    int voteCount = voteRepository.countByMenu(menu);
+                    int voteCount = Math.toIntExact(voteRepository.countByMenu(menu)); // Long -> int 변환
                     boolean votedByMe = user != null && voteRepository.existsByMenuAndUser(menu, user);
                     return new MenuResponseDto(menu, voteCount, votedByMe);
                 })
                 .collect(Collectors.toList());
+
 
         this.totalVotes = this.menus.stream().mapToInt(MenuResponseDto::getVoteCount).sum();
     }
